flask-server/requirements.txt
tinydb~=4.8.0
ibapi~=9.81.1
pycryptodome~=3.20.0
flask~=3.0.1
requests~=2.31.0
websocket-client~=1.7.0